back/requirements.txt
flask
flask-cors
diffusers
torch
transformers
celery
pillow
redis
accelerate
gunicorn
xformers
opencv-python
onnxruntime